crash.types: introduce new submodule for common types
This moves the list handling to crash.types.list and accompanying
infrastructure to crash.types.util so that it can be used elsewhere.
