+++ crash/types/__init__.py
@@ -1,0 +1,2 @@
+#!/usr/bin/env python
+# vim:set shiftwidth=4 softtabstop=4 expandtab textwidth=79:
